+++ cmd/init.go
@@ -15,7 +15,6 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/sveltinio/prompti/input"
-	"github.com/sveltinio/sveltin/common"
 	"github.com/sveltinio/sveltin/config"
 	"github.com/sveltinio/sveltin/helpers"
 	"github.com/sveltinio/sveltin/helpers/factory"
@@ -27,6 +26,7 @@ import (
 	"github.com/sveltinio/sveltin/internal/shell"
 	"github.com/sveltinio/sveltin/internal/tpltypes"
 	"github.com/sveltinio/sveltin/resources"
+	"github.com/sveltinio/sveltin/tui/feedbacks"
 	"github.com/sveltinio/sveltin/tui/prompts"
 	"github.com/sveltinio/sveltin/utils"
 )
@@ -190,7 +190,7 @@ func InitCmdRun(cmd *cobra.Command, args []string) {
 
 	cfg.log.Success("Done\n")
 
-	projectConfigSummary := &common.UserProjectConfig{
+	projectConfigSummary := &config.ProjectConfig{
 		ProjectName:   projectName,
 		CSSLibName:    cssLibName,
 		ThemeName:     themeSelection,
@@ -199,9 +199,9 @@ func InitCmdRun(cmd *cobra.Command, args []string) {
 
 	// NEXT STEPS
 	if themeData.ID != tpltypes.ExistingTheme {
-		common.PrintNextStepsHelperForNewProject(projectConfigSummary)
+		feedbacks.ShowNewProjectNextStepsHelpMessage(projectConfigSummary)
 	} else {
-		common.PrintNextStepsHelperForNewProjectWithExistingTheme(projectConfigSummary)
+		feedbacks.ShowNewProjectWithExistingThemeNextStepsHelpMessage(projectConfigSummary)
 	}
 
 }
