+++ config/project.go
@@ -1,0 +1,9 @@
+package config
+
+// ProjectConfig represents the user selections when creating a new sveltin project.
+type ProjectConfig struct {
+	ProjectName   string
+	CSSLibName    string
+	ThemeName     string
+	NPMClientName string
+}
